♻️ Refactor: postDetail을 post 파일로 변경
