+++ src/pages/PostPage.jsx
@@ -10,7 +10,7 @@ import BottomSheetModal from '../layouts/Modal/BottomSheetModal';
 import BottomSheetContent from '../layouts/Modal/BottomSheetContent';
 import ConfirmModal from '../layouts//Modal/ConfirmModal';
 import useModal from '../hooks/useModal';
-import { postDetailAPI } from '../api/apis/postDetail';
+import { postDetailAPI } from '../api/apis/post';
 
 const Main = styled.main`
   margin-top: 48px;
